src-python/pyproject.toml
[project]
name = "luna-backend"
version = "0.1.0"
description = "Luna backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "openai-whisper>=20231117",
    "torch>=2.2.0,<2.3.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "pygame>=2.5.0",
    "soundfile>=0.12.0",
    "aiohttp>=3.9.0",
]
